workaround parser throwing on unencoded `+` in querystring
`querystring` converts a literal `+` into a space when parsing
a url.  See joyent/node#8420 for background.

This is an issue for requests made from the address bar because
browsers don't encode `+` which is used by `sort` and `fields`.

This commit makes the `sort` and `fields` parsers accept a space
in place of a `+` at the start of each field.  It also adds an
option to turn off this behavior which would cause the parser
to throw an exception as it would have previously.